requirements.txt
torch
gymnasium
gymnasium[classic-control]
numpy
matplotlib
pyyaml
